model/requirements.txt
blinker==1.9.0
click==8.1.8
filelock==3.16.1
Flask==3.1.0
fsspec==2024.12.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.23.0
packaging==24.2
pillow==11.0.0
sympy==1.13.3
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
Werkzeug==3.1.3
